+++ plugins/hh-garcon/src/main/kotlin/ru/hh/plugins/garcon/extensions/psi/XmlTagExt.kt
@@ -4,13 +4,13 @@ import com.android.SdkConstants
 import com.android.tools.idea.util.androidFacet
 import com.intellij.psi.PsiClass
 import com.intellij.psi.xml.XmlTag
-import org.jetbrains.android.facet.LayoutViewClassUtils
+import org.jetbrains.android.facet.findClassValidInXMLByName
 import org.jetbrains.kotlin.idea.util.module
 
 
 fun XmlTag.toPsiClass(): PsiClass? {
     return module?.androidFacet?.let { facet ->
-        LayoutViewClassUtils.findClassByTagName(
+        findClassValidInXMLByName(
             facet,
             localName,
             SdkConstants.CLASS_VIEW