+++ shared/core/utils/src/main/kotlin/ru/hh/plugins/PluginsConstants.kt
@@ -6,6 +6,7 @@ object PluginsConstants {
     const val KOTLIN_FILE_EXTENSION = ".kt"
     const val ANDROID_LAYOUT_RES_DIR_NAME = "layout"
     const val BUILD_GRADLE_DEPENDENCIES_BLOCK_NAME = "dependencies"
+    const val BUILD_GRADLE_PLUGINS_BLOCK_NAME = "plugins"
 
     const val YAML_FILE_EXTENSION = ".yaml"
     val YAML_FILES_FILTER_EXTENSION = YAML_FILE_EXTENSION.removePrefix(".")